Revert '.md' extension replacement in relative links of versioned docs
Docusaurus V1 has had a bug for quite a long time related to the build
parameter named '--skip-next-release'. When this param is used for
building the site, the relative links in versioned docs do not change
the extension from '.md' to '.html' in the generated html files. This
produces site with invalid links.

To address the above mentioned issue, some logic was introduced so that
when versioned docs are created, the relative links within the source
files (markdown) are updated to have '.html' extension. This is not a
proper fix because when a docs contributor needs to update contents,
they may not even know that in versioned docs, links need to be written
with '.html' extension, instead of '.md' extension, which is the regular
method. Obviously, this introduces inconsistency also.

Removing this logic because this should not be handled by this "snapshot
version" creation. For ourselves, we've addressed the above bug in a
different manner.