Add GPL3 license, upgrade dependencies, and binary name
The prjoect did not have any license set when it was developed in-house.
Added GPL3 license after consulation internally.

Also updated all 3rd party dependencies to be on their latest version
so that there are no vulnerabilities. In addition, renamed the project
or binary from 'docusaurus-snapshotVersion' to
'docusaurus-snapshot-version'. This seems to be the proper convention
for module/package names.